Builds again...

+++ trunk/engine/ovr012.cs
@@ -80,7 +80,7 @@ namespace engine
             seg040.init_dax_block(out gbl.cursor_bkup, 0, 1, 1, 8);
             seg040.init_dax_block(out gbl.cursor, 0, 1, 1, 8);
 
-            seg051.FillChar(0xf, (ushort)gbl.cursor.bpp, gbl.cursor.data);
+            seg051.FillChar(0xf, gbl.cursor.bpp, gbl.cursor.data);
 
             gbl.symbol_8x8_set = new DaxBlock[5];
             gbl.symbol_8x8_set[0] = null;
